requirements.txt
pyyaml>=3.1.0
docopt
schema
